requirements.txt
langchain==0.3.3
langchain-community==0.3.2
langchain-ollama==0.2.0
langchain-qdrant==0.1.4
langchain-text-splitters==0.3.0

llama-index==0.11.17
llama-index-core==0.11.17
llama-index-embeddings-ollama==0.3.1
llama-index-llms-ollama==0.3.4
llama-index-storage-docstore-redis==0.2.0
llama-index-vector-stores-qdrant==0.3.1

qdrant-client==1.12.0
sentence-transformers==3.2.0
ollama==0.3.3

python-dotenv==1.0.1
